refactor: fix nits
