+++ extensions/algebra/rvr/ffi/modular/Cargo.toml
@@ -8,7 +8,7 @@ homepage.workspace = true
 repository.workspace = true
 
 [lib]
-crate-type = ["staticlib", "rlib"]
+crate-type = ["staticlib"]
 
 [dependencies]
 rvr-openvm-ext-algebra-ffi-common.workspace = true
